完善子类型的添加功能

+++ src/main/java/blog/controller/articaltype/ArticalTypeIndexController.java
@@ -36,7 +36,7 @@ public class ArticalTypeIndexController extends HttpServlet {
 		if(null != queryId)
 		{
 			type = new ArticalType();
-			type.setId(Short.valueOf(queryId));
+			type.setId(queryId);
 		}
 		
 		ArticalTypeDao typeDao = new ArticalTypeDaoImpl();
